Moving Github To Gitlab

+++ app/Http/Controllers/Shop/CategoryController.php
@@ -1,17 +1,25 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Shop;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Models\Category;
 use Illuminate\Support\Facades\Validator;
 
 class CategoryController extends Controller
 {
+
+
     public function show(Request $req)
     {
-        return Category::get();
+        $user = Auth::user();
+        $categories = Category::where('shop_id',$user->shop_id)->get();
+        return response()->json([
+            'status'=>true,
+            'categories'=>$categories
+        ]);
     }
     public function create(Request $req)
     {
@@ -28,15 +36,17 @@ class CategoryController extends Controller
             $category->name = $req->name;
             $category->remark = $req->remark;
             $category->shop_id = $req->shop_id;
-            if($shop->save()){
+            if($category->save()){
                 return response()->json(['status'=>true,"Category created successfully."]);
             }else {
                 return response()->json(['status'=>true,"Category can't created!"]);
             }
         }
     }
+
     public function update(Request $req)
     {
+        $user = Auth::user();
         $validator = Validator::make($req->all(), [
             'id'=> 'required',
             'name' => 'required|string',
@@ -47,35 +57,51 @@ class CategoryController extends Controller
         if ($validator->fails()) {
             return response()->json(['error' => $validator->errors()], 400);
         }else {
-            $category = Category::find($req->id);
-            $category->name = $req->name;
-            $category->shop_id = $req->shop_id;
-            $category->remark = $req->remark;
-            if($category->update()){
-                return response()->json(['status'=>true,"Category updated successfully."]);
-            }else {
-                return response()->json(['status'=>true,"Category can't updated!"]);
+            if($user->shop_id == $req->shop_id)
+            {
+                $category = Category::find($req->id);
+                $category->name = $req->name;
+                $category->remark = $req->remark;
+                if($category->update()){
+                    return response()->json(['status'=>true,"Category updated successfully."]);
+                }else {
+                    return response()->json(['status'=>true,"Category can't updated!"]);
+                }
             }
         }
     }
+
     public function delete(Request $req)
     {
-        $category = Category::find($req->id);
-        if($category->delete())
+        $user = Auth::user();
+        
+        if($user->shop_id == $req->shop_id)
         {
-            return response()->json(['status'=>true,'Message'=>"Category Deleted!"]);
-        }
-        else
-        {
+            $category = Category::find($req->id);
+            if($category->delete())
+            {
+                return response()->json(['status'=>true,'Message'=>"Category Deleted!"]);
+            }
+            else
+            {
+                return response()->json(['status'=>true,'Message'=>"Category can't delete!, Try Again"]);
+            }
+        }else {
             return response()->json(['status'=>true,'Message'=>"Category can't delete!, Try Again"]);
         }
-
     }
+
     public function restore(Request $req)
     {
-        $category = Category::withTrashed()->find($req->id);
-        if($category->restore()){
-            return response()->json(['status'=>true,"Category restored."]);
+        $user = Auth::user();
+        if($user->shop_id == $req->shop_id)
+        {
+            $category = Category::withTrashed()->find($req->id);
+            if($category->restore()){
+                return response()->json(['status'=>true,"Category restored."]);
+            }else {
+                return response()->json(['status'=>true,"Category can't restore!"]);
+            }
         }else {
             return response()->json(['status'=>true,"Category can't restore!"]);
         }
